+++ hits.c
@@ -6,7 +6,7 @@
 /*   By: auhoris <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/03 17:16:07 by auhoris           #+#    #+#             */
-/*   Updated: 2021/03/15 19:41:55 by auhoris          ###   ########.fr       */
+/*   Updated: 2021/03/15 21:28:19 by auhoris          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ t_floatp	horizontal_hit(t_cub *all, t_floatp fp, float pl_a)
 	dir = sinf(pl_a) > 0 ? 1 : -1;
 	hit.x = fp.x + dir * fabsf((float)y - fp.y) / tanf(pl_a);
 	hit.y = fp.y + dir * fabsf((float)y - fp.y);
-	while (hit.y > 0.0 && (int)hit.x < ft_strlen(all->map[(int)fabsf(hit.y)])
+	while (hit.y > 0.0 && (int)hit.x < (int)ft_strlen(all->map[(int)fabsf(hit.y)])
 			&& (int)hit.x > first_one(all->map[(int)(hit.y)]))
 	{
 		if (all->map[(int)(hit.y + dir * 0.00001)][(int)hit.x] == '1')
